apps/api/requirements.txt
eval_type_backport==0.3.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.4
httpx==0.27.0
python-multipart==0.0.9
boto3==1.34.129
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
anthropic==0.83.0
pytest==8.4.2
pytest-asyncio==1.2.0
scikit-learn==1.6.1
rapidfuzz==3.13.0
python-docx==1.1.0
